+++ Assets/Components/SpawnDelayComp.cs
@@ -6,12 +6,12 @@ using Unity.Entities;
 public struct SpawnDelayComp : IComponentData
 {
     public Vector2 movementVector; // direction and speed simultaneously
-    public int timer; //Amount of time delayed
+    public float timer; //Amount of time delayed
     public float radius;
     public byte mask;
     public float extraScale;
 
-    public SpawnDelayComp(Vector2 movementVector, int timer, float radius, byte mask, float extraScale)
+    public SpawnDelayComp(Vector2 movementVector, float timer, float radius, byte mask, float extraScale)
     {
         this.movementVector = movementVector;
         this.timer = timer;
